CustomAnalzyer, removes Wikipedia Markup and more
CustomAnalyzer, use it in WikipediaIR.java by replacing the instances of
EnglishAnalyzer with CustomAnalyzer.

Basically it is the EnglishAnalzyer, but now it removes Wikipedia Markup
using the WikipediaTokenizer and replaces tokens that match a given
regular expression with blanks (which are not indexed I am pretty sure)
using the PatternReplaceFilter. Right now, any token that ends in a
".com", ".org", ".jpg", or ".png" is replaced with a blank. The regular
expression can be extended to remove more tokens.

Not 100% sure it works as intended.

Wierd issues:

CustomAnalzyer without the PatternReplaceFilter create "out.txt" that
was bigger then the one created with EnglishAnalzyer?

CustommAnalzyer with PatternReplaceFilter creates "out.txt" that is
smaller but has more entries?

Lucene works with utf-8 and the xml documents are utf-8, not sure why
there are "?", tracked a term with a "?" in it and identified that
occurence of "?" as a  "ł"